Move .github/scripts/*.py to tools/
